fix bug #10 and add new hookpoint
